chore: move page inside layout and relocate component folder
