+++ src/components/home/letter-count-box.tsx
